updated Index.jsx
