+++ glaring-client/src/components/Event/Index.jsx
@@ -1,16 +1,48 @@
-/**@jsx jsx */
 import React from "react";
-import { css, jsx } from "@emotion/core";
+import "./Index.scss";
+import Button from "../Button";
+// import Axios from 'axios';
 
-export default function Event(props) {
-  const { props } = props;
+export default function Event() {
+  function accepted() {
+    console.log("Accepted invite");
+  }
+
+  function decline() {
+    console.log("Declined invite");
+  }
 
   return (
-    <div
-      css={css`
-        height: 200px;
-        width: 200px;
-      `}
-    ></div>
+    <div className="mom">
+      <main className="event">
+        <div className="event__top">
+          <img
+            className="event__image"
+            src="https://www.sciencenews.org/wp-content/uploads/2019/07/071019_MT_poker-ai_feat.jpg"
+          />
+          <div className="event__content">
+            <div className="event__title">EVENT TITLE</div>
+            <div className="event__description">
+              Event Descrtipion goes here
+            </div>
+          </div>
+        </div>
+        <div className="event__attend">
+          <div className="event__attend--list">123456789</div>
+          <div className="event__attend--reso">
+            <Button confirm onClick={accepted}>
+              ✓
+            </Button>
+            <Button danger onClick={decline}>
+              ✗
+            </Button>
+          </div>
+        </div>
+      </main>
+      <aside className="event__TIME">
+        <p>TI</p>
+        <p>ME</p>
+      </aside>
+    </div>
   );
 }
